Online Resume Builder
Welcome to the Online Resume Builder repository! This project was developed as part of my internship at CodeClause, where I served as a Java Development Intern.

Project Overview:

The Online Resume Builder is a powerful and versatile tool designed to simplify the process of creating professional resumes for developers and tech professionals. Built using Java, this web-based application leverages.

Contributions:

During my internship at CodeClause, I played a pivotal role in the development of this project. I contributed to the Java backend development